trip_review_service/requirements.txt
uvicorn
fastapi
python-dotenv
aio-pika
pytest
pytest-asyncio
pymysql
sqlalchemy
cryptography
PyJWT>=2.0.0
